Added another tower type
Added a lazer tower, drew another map. Made an upgrade button for gun
towers to machine gun towers. In it's current state, the game is
semi-playable in that you need a mix of towers to defend against the
waves of enemies.

+++ glfw-master/glfw_1/inc/Enemy.h
@@ -20,8 +20,8 @@ public:
 	void setSpeed(double speed);
 	double getSpeed();
 	enemyType getClass();
-	void setHp(int hp);
-	int GetHp();
+	void setHp(float hp);
+	float GetHp();
 	
 	//gives the enemy a new path
 	void setPath (EnemyMoveList path);
@@ -39,7 +39,7 @@ private:
 	enemyType m_Class;
 	double m_Speed;
 	EnemyMoveList m_Path;
-	int m_Health;
+	float m_Health;
 };
 
 // A linked list for the enemy path
